compiler warnings, sfml project platform changed to Windows XP

+++ 0.83/Omnibot/Common/MapGoal.cpp
@@ -1612,7 +1612,7 @@ bool MapGoal::SaveToTable(gmMachine *_machine, gmGCRoot<gmTableObject> &_savetab
 	if(m_Version!=MapGoalVersion) GoalTable->Set(_machine, "Version", gmVariable(m_Version));
 	GoalTable->Set(_machine,"GoalType",GetGoalType().c_str());
 	//GoalTable->Set(_machine, "Name", GetName().c_str());
-	int sz = m_Name.length() - m_GoalType.length() - 1;
+	int sz = (int)(m_Name.length() - m_GoalType.length() - 1);
 	if(sz <= 0 || m_Name.compare(m_GoalType.length()+1, sz, m_TagName)) 
 		GoalTable->Set(_machine, "TagName", GetTagName().c_str());
